pyproject.toml
[tool.poetry]
name = "python-queue-receiver"
version = "0.1.0"
description = ""
authors = ["Devlin Duldulao <webmasterdevlin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.3"
asyncio = "^3.4.3"
azure-servicebus = "^7.12.2"
python-dotenv = "^1.0.1"
azure-identity = "^1.17.1"
azure-keyvault-secrets = "^4.8.0"
pika = "^1.3.2"
common = {path = "../common"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

